+++ kong/credentials_service.go
@@ -7,19 +7,24 @@ import (
 	"reflect"
 )
 
+type credentialOptions struct {
+	credential interface{}
+	skipHash   bool
+}
+
 // abstractCredentialService handles credentials in Kong.
 type abstractCredentialService interface {
 	// Create creates a credential in Kong of type credType.
 	Create(ctx context.Context, credType string, consumerUsernameOrID *string,
-		credential interface{}) (json.RawMessage, error)
-	// Get fetches a credential of credType with consumerIdentifier and credIdentifier from Kong.
+		credentialOptions credentialOptions) (json.RawMessage, error)
+	// Get fetches a credential of credType with credIdentifier from Kong.
 	Get(ctx context.Context, credType string, consumerUsernameOrID *string,
 		credIdentifier *string) (json.RawMessage, error)
 	// GetByID fetches a credential of credType with ID from Kong.
 	GetByID(ctx context.Context, credType string, credIdentifier *string) (json.RawMessage, error)
 	// Update updates credential in Kong
 	Update(ctx context.Context, credType string, consumerUsernameOrID *string,
-		credential interface{}) (json.RawMessage, error)
+		credentialOptions credentialOptions) (json.RawMessage, error)
 	// Delete deletes a credential in Kong
 	Delete(ctx context.Context, credType string, consumerUsernameOrID, credIdentifier *string) error
 }
@@ -47,13 +52,18 @@ var credPathAsParentResource = map[string]string{
 	"mtls-auth":  "mtls-auths",
 }
 
+type queryStruct struct {
+	// Skip hashing of password when creating/updating basic-auth credentials.
+	SkipHash string `url:"skip_hash,omitempty"`
+}
+
 // Create creates a credential in Kong of type credType.
 // If an ID is specified in the credential, it will be used to
 // create a credential in Kong, otherwise an ID
 // is auto-generated.
 func (s *credentialService) Create(ctx context.Context, credType string,
 	consumerUsernameOrID *string,
-	credential interface{},
+	credentialOptions credentialOptions,
 ) (json.RawMessage, error) {
 	if isEmptyString(consumerUsernameOrID) {
 		return nil, fmt.Errorf("consumerUsernameOrID cannot be nil")
@@ -65,8 +75,8 @@ func (s *credentialService) Create(ctx context.Context, credType string,
 	}
 	endpoint := "/consumers/" + *consumerUsernameOrID + "/" + subPath
 	method := "POST"
-	if credential != nil {
-		if id, ok := credential.(id); ok {
+	if credentialOptions.credential != nil {
+		if id, ok := credentialOptions.credential.(id); ok {
 			if !reflect.ValueOf(id).IsNil() {
 				uuid := id.id()
 				if !isEmptyString(uuid) {
@@ -77,7 +87,15 @@ func (s *credentialService) Create(ctx context.Context, credType string,
 		}
 	}
 
-	req, err := s.client.NewRequest(method, endpoint, nil, credential)
+	var q *queryStruct
+
+	if credType == "basic-auth" && credentialOptions.skipHash {
+		q = &queryStruct{
+			SkipHash: "true",
+		}
+	}
+
+	req, err := s.client.NewRequest(method, endpoint, q, credentialOptions.credential)
 	if err != nil {
 		return nil, err
 	}
@@ -150,7 +168,7 @@ func (s *credentialService) GetByID(ctx context.Context, credType string, credId
 // Update updates credential in Kong
 func (s *credentialService) Update(ctx context.Context, credType string,
 	consumerUsernameOrID *string,
-	credential interface{},
+	credentialOptions credentialOptions,
 ) (json.RawMessage, error) {
 	if isEmptyString(consumerUsernameOrID) {
 		return nil, fmt.Errorf("consumerUsernameOrID cannot be nil")
@@ -164,8 +182,8 @@ func (s *credentialService) Update(ctx context.Context, credType string,
 	endpoint := "/consumers/" + *consumerUsernameOrID + "/" + subPath + "/"
 
 	credID := ""
-	if credential != nil {
-		if id, ok := credential.(id); ok {
+	if credentialOptions.credential != nil {
+		if id, ok := credentialOptions.credential.(id); ok {
 			if !reflect.ValueOf(id).IsNil() {
 				uuid := id.id()
 				if !isEmptyString(uuid) {
@@ -180,7 +198,14 @@ func (s *credentialService) Update(ctx context.Context, credType string,
 
 	endpoint = endpoint + credID
 
-	req, err := s.client.NewRequest("PATCH", endpoint, nil, credential)
+	var q *queryStruct
+
+	if credType == "basic-auth" && credentialOptions.skipHash {
+		q = &queryStruct{
+			SkipHash: "true",
+		}
+	}
+	req, err := s.client.NewRequest("PATCH", endpoint, q, credentialOptions.credential)
 	if err != nil {
 		return nil, err
 	}