+++ kong/jwt_auth_service.go
@@ -37,7 +37,9 @@ func (s *JWTAuthService) Create(ctx context.Context,
 	consumerUsernameOrID *string, jwtAuth *JWTAuth,
 ) (*JWTAuth, error) {
 	cred, err := s.client.credentials.Create(ctx, "jwt-auth",
-		consumerUsernameOrID, jwtAuth)
+		consumerUsernameOrID, credentialOptions{
+			credential: jwtAuth,
+		})
 	if err != nil {
 		return nil, err
 	}
@@ -94,7 +96,9 @@ func (s *JWTAuthService) Update(ctx context.Context,
 	consumerUsernameOrID *string, jwtAuth *JWTAuth,
 ) (*JWTAuth, error) {
 	cred, err := s.client.credentials.Update(ctx, "jwt-auth",
-		consumerUsernameOrID, jwtAuth)
+		consumerUsernameOrID, credentialOptions{
+			credential: jwtAuth,
+		})
 	if err != nil {
 		return nil, err
 	}