+++ kong/oauth2_auth_service.go
@@ -38,7 +38,9 @@ func (s *Oauth2Service) Create(ctx context.Context,
 	oauth2Cred *Oauth2Credential,
 ) (*Oauth2Credential, error) {
 	cred, err := s.client.credentials.Create(ctx, "oauth2",
-		consumerUsernameOrID, oauth2Cred)
+		consumerUsernameOrID, credentialOptions{
+			credential: oauth2Cred,
+		})
 	if err != nil {
 		return nil, err
 	}
@@ -96,7 +98,9 @@ func (s *Oauth2Service) Update(ctx context.Context,
 	oauth2Cred *Oauth2Credential,
 ) (*Oauth2Credential, error) {
 	cred, err := s.client.credentials.Update(ctx, "oauth2",
-		consumerUsernameOrID, oauth2Cred)
+		consumerUsernameOrID, credentialOptions{
+			credential: oauth2Cred,
+		})
 	if err != nil {
 		return nil, err
 	}